Changed some positions of code
Tried to move the else if statement containing the call to the blasterdeflect function up to under the if statement to check if the blade should be turned off.
Did this in hope that it would help on the sensor not being as reliable as I would want. It is unclear if this made any difference.