Adjustments for blade
Finally my led strips arrived! That means that I am ready to do adjustments for how the colors look inside the tube, as well as how long the animations should be with so many pixels. There is a lot of changes, so I won't cover them all. The biggest change is perhaps that the start animation is double as fast, because of the length of the blade.