Update NeoPixelSaberMainBranch.ino
Removed some hardcoding. The code now uses calculated variables for most of its code, so the blade length can be adjusted without changing lots of code

Added two new modes: BlueRedCrash and Spectrum. These are uniqe to the other modes in that they opperate with multiple colors at a time...